torch/triton_deployment/python_client/requirements.txt
pillow
opencv-python
opencv-python
opencv-contrib-python
torch
torchvision
torchsummary
tritonclient[all]
